refacotr: SRP를 지키기 위해 Project조회 서비스 로직 분리
