+++ sequence_member/src/main/java/sequence/sequence_member/project/service/ProjectBookmarkService.java
@@ -96,4 +96,16 @@ public class ProjectBookmarkService {
         }
     }
 
+    // 프로젝트 북마크 여부 확인 ( 북마크 되어있으면 true, 아니면 false, 로그인 안한 사용자는 false)
+    public boolean isBookmarked(Long projectId, CustomUserDetails customUserDetails) {
+        if (customUserDetails == null) {
+            return false;
+        }
+        MemberEntity member = memberRepository.findByUsernameAndIsDeletedFalse(customUserDetails.getUsername()).orElse(null);
+        if (member == null) {
+            return false;
+        }
+        return bookmarkRepository.existsByMemberIdAndProjectId(member.getId(), projectId);
+    }
+
 }
